test islands router navigation with page metadata
